Follow-up to commit 6166c36c23e745a6a98f31f0cdc05200c165f98f #1295
It turns out that commit broke running from an msys/cygwin/mingw
terminal.

I made the winconsole.py logic much more conservative and tolerant of
missing underlying libs or funny business.

Also now app will never abort if it fails to attach a console.. it will
just silently proceed.  It was a mistake to stop app if -v is not
successful.