Somewhat trivial nits: Windows console title + exceptions with WIF keys
- Windows console title should only really be set if running with -v
since setting the title creates the expectation of something to happen,
and without -v little to no output will be generated. So, fixed.

- Windows console title was being set but not unset on app exit. Fixed.

- Adjusted windows console title to be unambiguous.  The word 'Console'
means something else in our parlance (namely, the Console tab). Renamed
it to "Electron Cash - Verbose Output"

- bitcoin WIF key parsing raised BaseException, which is rather poor
form to raise such a general exception.  Narrowed it down to ValueError.
Calling code needs not be changed since presumably it was waiting for
BaseException anyway.