Console: Add warning dialog to protect against social engineering attacks
This adds a semi transparent overlay over the console window, which must be
acknowledged before the console can be interacted with. It warns the user
of the dangers of entering code and requires them to type a string into a
textbox to confirm they understand the risk. There is also a checkbox that
can be used to remember this confirmation so it doesn't have to be done
again for the loaded wallet.

closes #1297