Ledger: Fix importing libusb-1.0 after importing from PyQt5.Qt
PyQt5.Qt is a helper that scans through all PyQt5 modules to find the desired
import. This slowed down adding a hardware wallet as it loaded all the Qt
libraries. It also caused an issue where the PyQt5 stub of QtWebEngine was
loaded, which in turn loaded libcrypto.so.1.1 which caused a SIGSEGV in ld-linux
when loading libusb-1.0.so, at least on Arch Linux but likely on other platforms
as well.

Instead we just import from the correct module, PyQt5.QtWidgets