[Qt][Lib] Enable optional "Sign with Schnorr" in prefs
- Only allowed if the lib is present and working, and if the wallet type
is not multisig or hardware (we may enable for some limited harware
wallets in the future as they enable schnorr)

- Some minor refactorings to wallet.py

Note that only the GUI created TX's can ever have schnorr.  CashShuffle
is unaffected (it requires ECDSA).

The approach was to "pass down" the schnorr flag to functions that
create tx's, rather than consult wallet.storage.