QrReaderCameraDialog: Make async so camera can be loaded on Linux
On Linux the gstreamer plugins need a running message loop to load the
camera, but our code doesn't let it run after QCamera.load(). This
changes the dialog to an async model where the finished signal is emitted
when we are done scanning a QR code or got an error.