Fix #1375 - Receive tab should not present frozen addresses
Presumably when a user freezes an address -- they want to "take it out
of circulation".  This change ensures that the receive tab always
presents a completely unused and NOT frozen address.  The tab refreshes
itself every time it is shown -- so the user can freeze the receive
address in the tab and when they come back to the receive tab a new
address will be shown, etc.