Fix #1411: Qt bug on macOS; a repaint was missed for payto.setText
This only happens on macOS, and not consistently (read: even across EC
versions on the same system, it doesn't always happen).

There is no rhyme or reason to this bug, but the fact that it happens
means we needed a workaround.

The workaround is to issue repaint calls form the textChanged signal.

See issue #1411