Fix #1436 - Sometimes dependent PyQt5 packages are missing on Linux
In particular on Linux mint, QtSvg is missing even if they installed
the python3-pyqt5 distro package.

The real solution for users is to fully install pyqt5 via pip.

This commit simply displays a GUI messagebox telling the user to fully
install pyqt5 if the app detects QtSvg is missing at startup.

On Linux, it also tells the user what command to execute to do so.