PyInstaller: Rebuild bootloader from source using MinGW
For this we use a PyInstaller that has the windres bug fixed and we set
gcc to ignore the strncpy errors in the bootloader.

This results in a smaller binary that gets much less detections by
anti-virus programs. The binary dynamically links msvcrt.dll, but that
binary is available by default on every system we support.

See https://github.com/EchterAgo/pyinstaller/commit/d1cdd726d6a9edc70150d5302453fb90fdd09bf2