Windows: Fix to prevent excessive false positive AntiVirus detections
It turns out our false-positive woes started when we switched to
PyInstaller 3.4 release.  This version gets lots of false positives
becaue of the bootloader, runw.exe.  Sadly, many phishing and malware
authors probably use PyInstaller.

The fix was to recompile the windows 32 bit (and 64-bit, although we
don't currently use it) bootloader using a clean Windows 10 system with
MSVC 2015 via the `python ./waf all` as recommended by the pyinstaller
docs.  The generateed .EXE has a different hash and differs enough from
the canonical runw.exe shipped with pyinstaller release that we are now
down to 1 minor detection versus the 8-15 we would normally get with
vanilla PyInstaller.

We download this patched pyinstaller from our github repo and verify its
hash.  It is identical to the PyInstaller-3.4 tarball from the official
pip source, with only the .exe bootloader files differing.

We had to do this because the false positives were giving our users a
bad time.

Special thanks to Axel Gembe for helping to figure this out.