Localization: Default to system language and improve dialog
When the user has not chosen a language we now use the system
language or its best match by default. This takes care of selecting
for example es_AR as the preferred match for Spanish.

In the dialog we now show the language name translated to that
language and we include the language code in monospace as a prefix
to help sorting.

We move the General tab to the first position of the settings
dialog which makes it slightly easier to find the language
setting in case the UI is in the wrong language.

Finally we forgot to include 8 languages that are on Crowdin but
not in the GUI (Bulgarian, Mexican, Farsi, Korean, Swedish,
Turkish, Ukranian and Traditional Chinese)