Minor nit in Qt AddressList: Take cheap copy of internal wallet list.
We ended up iterating over a list that was an internal list for the
change and receiving address path in AddressList.  These lists are grown
in another thread for deterministic wallets. It's generally poor form to
iterate over data structures that another thread may modify.

While not catastrophic in Python since its threading semantics have guaranteed
atomicity on very high-level operations, it was still poor form.

So.. This is more of a nit-commit than anything.

The performane cost of copying lists in python via the list(other) constructor
is extremely cheap. I suspect internally CPython returns a handle to the
same list, which then employs copy-on-write semantics should either of
the two lists be modified later.