AppImage: Bundle fontconfig / freetype and load if needed
The Ubuntu 16.04 freetype / fontconfig does not seem to properly
support color emojis. This patch installs a newer freetype / fontconfig
combo from the Ubuntu 18.04 repo and includes its dependencies of
libpng 16 and zlib 1.2.11.

On startup of the AppImage, we test if the systems freetype is newer
than the one in the AppImage, and if it is not, we add our fonts folder,
which is where we put freetype / fontconfig, to the library path.