Wallet: Improve performance of is_mine() check from O(N) to O(logN)
This check was using the address list each time it was called to
effectively test if an address is in the wallet. List __contains__
lookups are O(N) for the size of the list.

We instead maintain an address set as a cache for the is_mine check and
this means walle.is_mine() checks are now O(logN).

While this doesn't matter for wallets with a small set of addresses, for
large wallets like CashShuffle wallets (5k or 10k addresses are not
unheard of), this optimization makes a difference.

This is especially true becaue wallet.is_mine is called very frequently,
especially when starting up and/or synchronizing the wallet or rebuilding
its history.