Fix #1502 - Intermittent iOS Crash Reporter on bad/expired BIP70 Payment
This mainly happens because there is a race condition between when the
bad payment request variable is cleared with respect to the screen
refresh callbacks.

Since a payment request without an outputs attribute is invalid anyway,
it's safe to catch this exception and proceed as if the PR is not there.

The payment request gets cleared immediately after the "This is an
expired payment request" dialog is dismissed.