Enforce specific token type with Validator_SLP1
* token type 0x01 and 0x81 (NFT1 parent) use the same SLP1
validator, so we just need to make sure we don't allow transactions
to change between 0x01 and 0x81 in the same graph.  The default type
is 0x01 but can be specified in the Validator_SLP1 constructor using the
'enforced_token_type' parameter.

* added a new invalid state (4) associated to changing token type.  This
will allow us to debug issue specific to wallet implementations not honoring
the correct SLP message token_type.