Fix #1521: Pay To edit can have the line disappear
See issue #1521. Basically this is some sort of quirk in Qt where if you
select-drag with mouse, the scrollbar bounces off the edge of the payto
edit and the contents are no longer visible.

The fix was to not allow the scrollbar to go past minimum() on
single-liners.

I suspect this bug is triggered by the funny business in update_size()
in this PayToEdit class, but that code seems fragile so I will not
investigate redoing that. Instead this workaround works well.