Lib: DoS protection from malicious server
- Limit the banner message to 16kb. Previously it was possible for a
server to send a huge block of emoji characters which could cause
clients to hang for many minutes at a time as they struggle to render
them (in particular Linux seems to be slower to render emoji
characters).

- Limit the message size from the server to 32MB. Previously it was
possible for a malicious server to continually send garbage data,
thereby crashing clients by filling their memory eventually until the
process segfaults.  Now, we limit the receive buffer to 32MB and if this
limit is exceeded, the connection is automatically closed.