Qt UI: Fixups to wallet search facility (Ctrl-F)
- The search box was all the way on the right and quite often if running
CashShuffle, would be scrunched up and tiny and unusable because the
balance label would take up all the space. This has been fixed. The balance
label is now hidden while the search box is up. Closes #1493.

- It was not immediately obvious what the search box DOES. Placeholder
text has been added instructing the user that it searches the wallet and
also how to hide the search box (Ctrl-F again).

- Previously the Ctrl-F search was quirky. It only applied to the
current tab. If you switched tabs the search box showed some search
text, but the new tab would not have the search applied. This was
misleading and not ideal. Now the search is applied to all tabs
simultaneously.

- If hiding the search box and showing it again, you would see the old
search text in the box, but the old search filter would not be
re-applied for you. This was misleading and strange. This has been
addressed.  Hitting Ctrl-F reveals the search box AND also applies the
old filter.