Send tab: (Qt) Updated the "Pay to" Help Label text.
It is much clearer about what the pay-to field supports.  We may also
add a "?" tool button to the payto field for this text additionally, as
it's not immediately obvious that users normally see the "Help Labels"
(they are hover links).

Apologies to translators, as this string has HTML embedded in it -- but
there was no other way to really format it in a pretty way other than
doing this.

It's hoped that by providing it as 1 long string to translators, they
get a bit more context.. and can work their way around the HTML bits.