Windows & Linux - FreeType & fontconfig logic tweaks + prefs options (#1544)
* Windows Qt: Added preferences option for enabling/disabling FreeType

The default to this option is True for Windows < Win10, False otherwise.

The preferences option is only presented on Windows (sys.platform ==
win32 or cygwin).  On other platforms it does not appear and its config
key has no effect.

This addresses the fact that for some users FreeType fonts are a good
idea, but for others perhaps they would like finer grained control.

FreeType in general is good but it does lack subpixel rendering so for
some users the tradeoff there may not warrant using it.

* Linux: Added fontconfig override GUI setting for more custom control

Perhaps a dangerous setting -- but not offering it perhaps also has its
own drawbacks.

* Added paranoia in getter method to enforce bool no matter what config has

* Fixup for Windows 10 detection -- use platform.win32_ver

* Handle windows 8.1 in freetype detect code

* Moved the _use_fontconfig and _use_freetype config methods to gui_object

* Fix