Build: Run docker as current user
Right now commands in docker run as root, which will leave
root-owned output in the source directory. This changes docker to
run the commands as the currently logged in user instead. This
removes the need for root priviledges to remove the fresh clone
directory.

This also moves the location of the GPG keyring into the fresh
clone directory as otherwise GPG would try to create it in the images
root directory.