Build: Add ability to skip tagging the PyInstaller bootloader
This can be useful in some cases, for example when you do build system
work, like in the following commits, where you want to make sure the
build output doesn't change. Just prefix your build command with
PYI_SKIP_TAG=1 and it will not be tagged.

This also fixes the message about detached head in the PyInstaller build
and removes the version number included in the build.