Reverted wine build to use less of host FS -- was too slow on macos
Added a /homedir in the docker image which is owned by UID:GID of
current user.

WINEPREFIX gets put in /homedir/wine64
Electron Cash dir (mounted to hostfs) get put in
/homedir/wine64/drive_c/electroncash

The Dockerfile makes sure to first create these directories and set
theor ownership to UID:GID.

More of the guest FS is now being used than in the build_detect_tty
branch and even more is used than on current master (tmp is now on guest
FS).

The build is fast again.