Qt bugfix: Closes #1554 - Window close before load_wallet called
Reworked the code a little bit to be more defensive against pathological
cases where wallet.thread may not be yet defined when closeEvent
arrives.

This can only really happen in pathological plugin circumstances,
however, but it pays to be resilient against this situation.

See #1554.

This commit makes the code that cleans up the window be more careful
when accessing objects for cleanup (such as wallet.thread which may not
be there always).

It also now calls super().closeEvent() as per Qt spec (it did not before
and this is out of spec for Qt).