Refactor: main_window add_token_type ->  wallet add_token_safe
The checks done in main_window.py's add_token_type() method really are
generic checks that could be re-used by other code (eg. daemon/CLI/RPC
calls).

So we refactored them into the wallet in a method called
'add_token_safe' which the main_window.py 'add_token_type' method
basically wraps.

This is a necessary step towards code re-use and also adding RPC/CLI and
some day other platform support for the SLP backend.

Note:

- We now strictly enforce ALL LOWERCASE on token id's. This shouldn't be a
problem with existing code.  But we do NOT want mixed-case or uppercase
hex strings to make it into the wallet's storage as that would create
chaos.

- This commit fixes a potential race condition which existed with the
token_types dict being modified as it was being iterated because it was
accessed without locks.  We instead take a snapshow copy of it when
iterating over it without locks.