Revert NFT validator back to a per-wallet validator
The shared validator has bugs. The same bugs are in master. Reading the
code it appears to clobber previous wallets, for one (thus it can't
really be shared).

Also -- if you open another wallet with the same NFT token the
parent/child relationship between the tokens is lost.

Keeping the validators per-wallet for NFT is a workaround for these
issues.