Added is_multi support to validator GraphContext* classes
If the GraphContext is constructed with is_multi=True, it will create 1
ValidationJobManager (eg 1 thread) per token id and tokens validate in
parallel.

If is_multi=False (the default for now, although this should probably
change) tokens validate in series and there is 1 ValidationJobManagr per
GraphContext instance.

This commit also fixes the kill_graph() implementation to work
correctly (previously it was ill-defined).