Validator: Parallelized slp_v1 (non-NFT) validator
- The validations now run in parallel, 1 thread per coin waiting to be
validated.
- As before, the validation context is shared across open wallets
- Wallet close, as before, stops the jobs related to that wallet. When
all jobs are stopped, the ValidationJobMgr is stopped and its threads
exit.
- When the token is validated, its job mgr thread is also stopped and
the thread exists.

This facility can be turned off and reverted back to serial by simply
constructing the GraphContext object with `is_parallel=False` as an
argument to __init__

Tested and works.  Speedup on restoring a wallet with many tokens for me
here is 3x over serial validation.