Daemon: Prevent infinite hangs when lockfile create_time is in the future
This should address all the extant issues users sometimes report with
the daemon lockfile causing Electron Cash to never start up.  99% of the
issues related to this were the create_time being in the future.

Previously the daemon/lockfile mechanism would hang forever if the
lockfile's create_time was in the future. This can happen if there was a
system clock adjustment since the last run of the program and the
lockfile remained present (which the previous code didn't even ensure
would be deleted in most normal GUI uses!)

Now, we:

- Ensure the lockfile is removed if created by us using atexit.
- We no longer tolerate "connection failed" + create_time in the future:
Obviously this means there was a clock adjustment.
- We also take steps to prevent infinite hangs due to other filesystem
funny business after > 5 retries.

This commit fixes #1240, #892, #1484