LabelSync Plugin: Prevent error window spam
When importing labels from file, the LabelSync plugin ends up
overloading the server with requests and eventually it spams the main
window with error dialogs as the requests time out.

We now prevent the error dialog spam by only showing at most 1 error
dialog window per main_window for labelsync.

Previously it was possible to get 100 or more windows you'd have to
dismiss, all on top of each other!

The underlying problem with the plugin however is that it needs to be
more resilient and also batch uploads to the server.

But for now, this workaround at least prevents pathological UX.