Localization: Make age and time_difference function translatable
This also adjust the cutoff for the "about 1 day / month / year"
text to 1.5 days / months / years, just like with hours.

It also fixes the since_date parameter to be converted from
timestamp just like from_date. This will happen only when
since_date is an integer or float to not break any code.