Lib: small nit in bitcoin.py
Fix for bitcoin.py base_decode:

This function would accept characters not in the base58 or base43 character
set and would produce garbage bytes output on strings outside the character
set. The function has been made more strict to raise ValueError when it
encounters bad characters in the character stream.  All functions that
call it (such as DecodeBase58Check) have been updated to catch this
exception and behave correctly.