Console: Make warning fit in default window size
Back we initially added the console warning, we didn't do
automatic high DPI scaling except on macOS. Enabling the scaling
caused the console warning to be too large for the default window
size on Windows and Linux.

This just reduces the padding / margins a bit, decreases the font
size and keeps the Wikipedia link in the same paragraph.