Revert MacOS build to PyQt5 5.11.x until Qt 5.13.1 is released.
The reason for this is because there are issues with PyQt5 5.12.x &
5.13.0. QPushButtons sometimes don't send click events (they get dropped
basically).  This has been fixed in 5.13.1 dev snapshot.

So, we revert to PyQt5 5.11.x which is less than ideal because:

Mojave doesn't work quite right on 5.11 (dark mode in particular is not
supported, as well as other misc. rendering quirks).  However, it's better
to use 5.11, with the Mojave caveats in mind, than 5.12 and 5.13.0 which
drop UI events if the main thread is busy.

We will re-enable Qt 5.13.x when PyQt5 5.13.1 is released (this works
much better on all platforms at the expense of dropping support for El
Capitan).

See Qt Bug:
http://bugreports.qt.io/browse/QTBUG-69687

Commit that fixes this:
https://code.qt.io/cgit/qt/qtbase.git/commit/?h=5.13.1&id=09a2a9bc4a4a5c9ef7849d3f30243cd8e2d389d1