Graph Search validation speed improvement (#81)
This PR includes adding a new class `GraphSearchManager` to fetch transactions from SLPDB and store them in a special transaction cache reserved for GS results.  `GraphSearchManager` lives as a singleton within the Token Type 0x01 GraphContext, and hooks in at the `fetch_hook` callback to kickoff new searches as needed.

GraphSearchManager manages a single thread with non-blocking queues to process the searches.  Per discussions w/ @cculianu we will need to reduce the complexity of this before we merge with mainline.

This feature can be enabled (as "opt-in" only) via a new tab in network_dialog call Tokens.