Add BIP38 Encryption to the GUI (#1594)
- Wallet -> Private Keys -> Export -> BIP38 Encrypted...; prompts
   user for passphrase, exports all keys encrypted with said passphrase.
- In Private key dialog (accessible from address right-click menu), added
   a "Encrypt BIP38..." button which prompts for passphrase and encrypts
   the displayed WIF key using BIP38

Also in this PR:

- GUI nits: Made all menubar menus follow the UI standard "Thing..."
   if "Thing" leads to a dialog box.

- Added password dialog box checkbox "Show" which shows the password
   the user is inputting. This is a standard GUI thing to do since some users
   prefer to see what they are typing. Off by default.

---

* Added BIP38 Private key export + nits

- BIP38 Private key export is in Wallet -> Private keys -> Export menu.
  (It prompts the user for a passphrase and is otherwise identical to
  the bare WIF export except for the fact that it produces BIP38 key
  output)
- UI Nits: Made all app menus use elipses properly "..." for menu
actions that lead to another dialog.
- Nits: Made traceback.print_exc print to stderr and also take into
account is_verbose app-wide flag.

* Qt: PasswordLayout nits

- Added "Show" checkbox to the password dialog to set echo mode on
QLineEdits to normal. Default unchecked.

- Misc. nits in PasswordLayout

* Qt GUI: Sign message warning text nit

The 'msg_sign' variable was class-level and was untranslated as a result
(language & i19n isn't set necessarily when the class is being evaluated by
Python). Moved this variable init insite the method implementation.
There was no good reason for it to be class-level since it was a very
rarely used variable.

Also made the sign_message callpath for multisig abort, not try and sign
the message with a bogus private key.

* Privae key dialog: Added "Encrypt BIP38..." button

This button allows the user to encrypt the displayed WIF key using BIP38.