daemon CLI: Made addrequest print some useful error messages
It was returning "False" when it couldn't create a new request with no
explanation as to why.  Added some error messages instructing the user
what the situation is.

Also refuse to try and "create_new_address" on non-HD wallet (where it
will crash), and instead return False but put some messages on the -v
log.