Minor nits: rm dead code; use tx.outputs() in favor of tx.get_outputs()
We probably should delete tx.get_outputs() since it's slower than
tx.outputs(), but plugins may use it so we'll leave it around.

But in the wallet we switched all calls to the slightly faster tx.outputs().

Also removed some dead code in histroy_list create_menu which was
calling get_wallet_delta() for no reason.