Wallet: Clean-up and fix code in add_transaction
The add_transaction code had a few minor tiny problems:

- It was unnecessarily slow (looking up the same values multiple times
in the various dicts it manipulated).  Redundant dict lookups have been
eliminated.

- When processing a transaction's inputs, it would compute various items
such as `ser` and `prevout_hash` and `prevout_n`, etc, unconditionally
even if it was not going to use that information at all 2 lines later.
The calculation of those values is only done now when it's clear they
will be used. This is a tiny CPU savings but perhaps can shave
milliseconds off synchronizations of large wallet histories.

- It kept self.txi and self.txo around for transaction inputs/outputs
with no relevance to wallet (these would get serialized as empty dicts
to the wallet file, bloating the wallet, and also bloating memory at
runtime by a small amount).

- It had a minor logical error when moving from pruned_txo -> self.txi
(the actual self.txi dict wasn't being updated potentially).

So with this commit, self.txi and self.txo never save irrelevant data to
the wallet file, saving space and memory.  This change should be
backward compatible with older installations (since they saved the data
but pretty much ignored it). Note that the previous space waste was not
significant (it was just a bunch of empty dicts per tx seen), but it
could potentially add up for thousands of transactions in the wallet's
history.

This commit also fixes some minor nits and adds some minor comments.